Auto stash before merge of "codex/add-windows-progress-bar-feature" and "origin/codex/add-windows-progress-bar-feature"
